+++ app/src/components/TxSubmit/SendTx.tsx
@@ -6,21 +6,16 @@ import type { ExtrinsicPayloadValue, ISubmittableResult } from '@polkadot/types/
 import type { HexString } from '@polkadot/util/types';
 import type { BuildTx } from './hooks/useBuildTx';
 
-import IconClock from '@/assets/svg/icon-clock.svg?react';
-import IconQuestion from '@/assets/svg/icon-question-fill.svg?react';
 import { CONNECT_ORIGIN } from '@/constants';
 import { addTxToast } from '@/hooks/useTxQueue';
 import { useAccountSource } from '@/wallet/useWallet';
 import { enableWallet } from '@/wallet/utils';
 import React, { useState } from 'react';
 
-import { addressToHex, sign, signAndSend, TxEvents, useApi } from '@mimir-wallet/polkadot-core';
+import { sign, signAndSend, TxEvents, useApi } from '@mimir-wallet/polkadot-core';
 import { service } from '@mimir-wallet/service';
-import { Alert, Button, Divider, Tooltip } from '@mimir-wallet/ui';
+import { Alert, Button } from '@mimir-wallet/ui';
 
-import AddressName from '../AddressName';
-import FormatBalance from '../FormatBalance';
-import LockItem, { LockContainer } from '../LockItem';
 import { toastError } from '../utils';
 import { useDryRunResult } from './hooks/useDryRunResult';
 
@@ -57,19 +52,9 @@ function SendTx({
 }) {
   const { api, network } = useApi();
   const [loading, setLoading] = useState(false);
-  const { txBundle, isLoading, error, hashSet, reserve, unreserve, delay } = buildTx;
-  const [enoughtState, setEnoughtState] = useState<Record<HexString, boolean | 'pending'>>({});
+  const { txBundle, isLoading, error, hashSet, delay } = buildTx;
   const source = useAccountSource(txBundle?.signer);
   const { data: dryRunResult } = useDryRunResult(txBundle);
-
-  const isEnought = Object.keys(reserve).reduce<boolean>(
-    (result, item) => result && !!enoughtState[addressToHex(item)],
-    true
-  );
-  const isEnoughtPending = Object.keys(reserve).reduce<boolean>(
-    (result, item) => result || enoughtState[addressToHex(item)] === 'pending',
-    false
-  );
 
   const onConfirm = async () => {
     let events: TxEvents = new TxEvents();
@@ -168,70 +153,7 @@ function SendTx({
 
   return (
     <>
-      {(Object.keys(reserve).length > 0 || Object.keys(unreserve).length > 0 || Object.keys(delay).length > 0) && (
-        <LockContainer>
-          {Object.entries(delay).map(([address, delay], index) => (
-            <div key={`delay-${address}-${index}`} className='flex items-center justify-between gap-[5px] sm:gap-2.5'>
-              <div className='flex items-center gap-[5px] sm:gap-2.5'>
-                <IconClock className='text-primary h-4 w-4 opacity-50' />
-                <p>Review window</p>
-                <Tooltip
-                  content='This transaction needs to be executed manually after review window ends.'
-                  closeDelay={0}
-                >
-                  <IconQuestion className='text-primary/40' />
-                </Tooltip>
-              </div>
-
-              <span>{delay.toString()} Blocks</span>
-            </div>
-          ))}
-
-          {Object.keys(delay).length > 0 && <Divider className='bg-primary/5' />}
-
-          {Object.entries(reserve).map(([address, { value }], index) => (
-            <LockItem
-              key={`lock-${address}-${index}`}
-              address={address}
-              isUnLock={false}
-              value={value}
-              tip={
-                <>
-                  <FormatBalance value={value} /> in{' '}
-                  <b>
-                    <AddressName value={address} />
-                  </b>{' '}
-                  will be reserved for initiate transaction.
-                </>
-              }
-              onEnoughtState={(address, isEnought) =>
-                setEnoughtState((state) => ({ ...state, [addressToHex(address)]: isEnought }))
-              }
-            />
-          ))}
-          {Object.entries(unreserve).map(([address, { value }], index) => (
-            <LockItem
-              key={`unlock-${address}-${index}`}
-              address={address}
-              isUnLock
-              value={value}
-              tip={
-                <>
-                  <FormatBalance value={value} /> in{' '}
-                  <b>
-                    <AddressName value={address} />
-                  </b>{' '}
-                  will be unreserved for execute transaction.
-                </>
-              }
-            />
-          ))}
-        </LockContainer>
-      )}
-
       {error ? <Alert color='danger' title={<span className='break-all'>{error.message}</span>} /> : null}
-
-      {!isEnought && !isEnoughtPending ? <Alert color='danger' title='Insufficient funds' /> : null}
 
       {Object.keys(delay).length > 0 ? (
         <Alert color='warning' title='This transaction can be executed after review window' />
@@ -245,9 +167,7 @@ function SendTx({
         color='primary'
         onPress={error ? undefined : onConfirm}
         isLoading={loading || isLoading}
-        isDisabled={
-          !txBundle?.signer || !!error || !isEnought || disabled || (dryRunResult ? !dryRunResult.success : false)
-        }
+        isDisabled={!txBundle?.signer || !!error || disabled || (dryRunResult ? !dryRunResult.success : false)}
       >
         Submit
       </Button>
