+++ app/src/params/Call/FunctionArgs.tsx
@@ -28,7 +28,7 @@ const FunctionArgs = forwardRef<HTMLDivElement | null, CallProps>(({ registry, c
   }, [registry, call]);
 
   return done ? (
-    <div ref={ref} className={mergeClasses('', className)}>
+    <div ref={ref} className={mergeClasses('flex flex-col gap-2.5', className)}>
       {args ? (
         args.length > 0 ? (
           args.map(([name, type], index) => (
