+++ app/src/params/Call/index.tsx
@@ -12,7 +12,6 @@ import CancelAsMulti from './CancelAsMulti';
 import FunctionArgs from './FunctionArgs';
 import SetIdentity from './SetIdentity';
 import TransferCall from './TransferCall';
-import { useRefState } from './withRenderState';
 
 type MatchType = 'exact' | 'prefix';
 
@@ -50,20 +49,11 @@ const componentDef: ComponentConfig[] = [
 ];
 
 const Call = forwardRef<HTMLDivElement | null, CallProps>((props, ref) => {
-  const {
-    registry,
-    call,
-    showFallback = false,
-    fallbackComponent: FallbackComponent = FunctionArgs as CallComponent,
-    ...restProps
-  } = props;
+  const { registry, call, showFallback = false, fallbackComponent: FallbackComponent = FunctionArgs } = props;
 
   const action = useMemo(() => (call ? findAction(registry, call)?.join('.') : null), [call, registry]);
   const domRef = useRef<HTMLDivElement>(null);
   const currentRef = ref || domRef;
-
-  // Track the render state of the current ref
-  const renderState = useRefState(currentRef);
 
   // Find matching component
   const Component = useMemo(() => {
@@ -83,20 +73,12 @@ const Call = forwardRef<HTMLDivElement | null, CallProps>((props, ref) => {
     return null;
   }, [action]);
 
-  if (!action) {
-    return null;
+  if (!action || !Component) {
+    return showFallback ? <FallbackComponent ref={currentRef} {...props} /> : null;
   }
 
   // If we have a matched component, render it with render state detection
-  return (
-    <>
-      {Component ? <Component ref={currentRef} {...props} /> : null}
-
-      {(!Component || renderState === 'empty') && showFallback ? (
-        <FallbackComponent ref={!Component ? currentRef : undefined} registry={registry} call={call} {...restProps} />
-      ) : null}
-    </>
-  );
+  return <Component ref={currentRef} {...props} />;
 });
 
 Call.displayName = 'Call';
